+++ source/games/exhumed/src/ramses.cpp
@@ -161,8 +161,8 @@ void InitSpiritHead()
 
 void DimSector(int nSector)
 {
-    short startwall = sector[nSector].wallptr;
-    short nWalls = sector[nSector].wallnum;
+    int startwall = sector[nSector].wallptr;
+    int nWalls = sector[nSector].wallnum;
 
     for (int i = 0; i < nWalls; i++)
     {
